+++ source/ai/StackAlphaBetaSearch.cpp
@@ -35,7 +35,7 @@ void StackAlphaBetaSearch::doSearch(const GameState & initialState, Move & move)
     _searchTimer.start();
     unsigned long long lastNodesSearched = 0;
 
-    for (; _currentMaxDepth <= _params.maxDepth(); _currentMaxDepth++)
+    for (; (int)_currentMaxDepth <= _params.maxDepth(); _currentMaxDepth++)
     {
         unsigned long long previousNodes = _results.nodesExpanded;
